Code for debugging
There were parse errors on brat, apparently because of the mismatch of
text span due to unicode character counting in the data and the default
ascii character counting in java.
I added a constant to suppress writing output onto files; now
triggering whether to write output into files can be set with a
property mydcoref.output.

Still working on counting with unicode character instead of ascii in
Java