Add option to output as .token
Add option to output as .token. The file includes information about
tokens and coreference chains just as you would find in Stanford XML
format. However, the token offsets are not given because encoding
complications, so the output must be processed with other program (will
be written in C#) to convert the files into brat standoff format.